:truck: move: auth 관련 파일 정리 #52
