--- src/main/java/com/ddd/oi/auth/dto/AuthResponseDTO.java
@@ -1,6 +1,0 @@
-package com.ddd.oi.auth.dto;
-
-import com.ddd.oi.user.domain.User;
-
-public record AuthResponseDTO(User user, String accessToken,String refreshToken,String oauthAccessToken) {
-}
