+++ src/main/java/com/ddd/oi/common/utils/GoogleUtil.java
@@ -1,11 +1,10 @@
 package com.ddd.oi.common.utils;
 
-import com.ddd.oi.auth.dto.GoogleDTO;
+import com.ddd.oi.auth.dto.profile.GoogleDTO;
 import com.ddd.oi.common.exception.OiException;
 import com.ddd.oi.common.response.ErrorCode;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
